+++ lib/NewWindow.d.ts
@@ -1,9 +1,0 @@
-import React from "react";
-import { BoxData } from "./DockData";
-interface Props {
-    panelData: BoxData;
-}
-export declare class NewWindow extends React.PureComponent<Props, any> {
-    render(): React.ReactNode;
-}
-export {};