+++ src/DockLayout.tsx
@@ -395,7 +395,7 @@ export class DockLayout extends DockPortalManager implements DockContext {
     }
 
     DragManager.addDragStateListener(this.onDragStateChange);
-    window.addEventListener('resize', this._onWindowResize);
+    globalThis.addEventListener?.('resize', this._onWindowResize);
   }
 
   /** @ignore */
@@ -571,7 +571,7 @@ export class DockLayout extends DockPortalManager implements DockContext {
 
   /** @ignore */
   componentWillUnmount(): void {
-    window.removeEventListener('resize', this._onWindowResize);
+    globalThis.removeEventListener?.('resize', this._onWindowResize);
     DragManager.removeDragStateListener(this.onDragStateChange);
     this._onWindowResize.cancel();
     this._isMounted = false;