+++ src/DockPanel.tsx
@@ -25,6 +25,12 @@ export class DockPanel extends React.PureComponent<Props, State> {
   _ref: HTMLDivElement;
   getRef = (r: HTMLDivElement) => {
     this._ref = r;
+    if (r) {
+      let {parent} = this.props.panelData;
+      if ((parent?.mode === 'float')) {
+        r.addEventListener('pointerdown', this.onFloatPointerDown, {capture: true, passive: true});
+      }
+    }
   };
 
   static _droppingPanel: DockPanel;
@@ -77,7 +83,7 @@ export class DockPanel extends React.PureComponent<Props, State> {
     let {panelData} = this.props;
     let {parent, x, y, z} = panelData;
     let dockId = this.context.getDockId();
-    if (parent && parent.mode === 'float') {
+    if (parent?.mode === 'float') {
       this._movingX = x;
       this._movingY = y;
       // hide the panel, but not create drag layer element
@@ -126,6 +132,18 @@ export class DockPanel extends React.PureComponent<Props, State> {
   _movingW: number;
   _movingH: number;
   _movingCorner: string;
+  onPanelCornerDragT = (e: DragState) => {
+    this.onPanelCornerDrag(e, 't');
+  };
+  onPanelCornerDragB = (e: DragState) => {
+    this.onPanelCornerDrag(e, 'b');
+  };
+  onPanelCornerDragL = (e: DragState) => {
+    this.onPanelCornerDrag(e, 'l');
+  };
+  onPanelCornerDragR = (e: DragState) => {
+    this.onPanelCornerDrag(e, 'r');
+  };
   onPanelCornerDragTL = (e: DragState) => {
     this.onPanelCornerDrag(e, 'tl');
   };
@@ -141,7 +159,7 @@ export class DockPanel extends React.PureComponent<Props, State> {
 
   onPanelCornerDrag(e: DragState, corner: string) {
     let {parent, x, y, w, h} = this.props.panelData;
-    if (parent && parent.mode === 'float') {
+    if (parent?.mode === 'float') {
       this._movingCorner = corner;
       this._movingX = x;
       this._movingY = y;
@@ -166,6 +184,24 @@ export class DockPanel extends React.PureComponent<Props, State> {
     }
 
     switch (this._movingCorner) {
+      case 't': {
+        panelData.y = this._movingY + dy;
+        panelData.h = this._movingH - dy;
+        break;
+      }
+      case 'b': {
+        panelData.h = this._movingH + dy;
+        break;
+      }
+      case 'l': {
+        panelData.x = this._movingX + dx;
+        panelData.w = this._movingW - dx;
+        break;
+      }
+      case 'r': {
+        panelData.w = this._movingW + dx;
+        break;
+      }
       case 'tl': {
         panelData.x = this._movingX + dx;
         panelData.w = this._movingW - dx;
@@ -249,11 +285,8 @@ export class DockPanel extends React.PureComponent<Props, State> {
     let isHBox = parent?.mode === 'horizontal';
     let isVBox = parent?.mode === 'vertical';
 
-    let pointerDownCallback = this.onFloatPointerDown;
     let onPanelHeaderDragStart = this.onPanelHeaderDragStart;
-    if (!isFloat || isMax) {
-      pointerDownCallback = null;
-    }
+
     if (isMax) {
       dropFromPanel = null;
       onPanelHeaderDragStart = null;
@@ -295,12 +328,23 @@ export class DockPanel extends React.PureComponent<Props, State> {
 
     return (
       <DragDropDiv getRef={this.getRef} className={cls} style={style} data-dockid={id}
-                   onMouseDownCapture={pointerDownCallback} onTouchStartCapture={pointerDownCallback}
                    onDragOverT={isFloat ? null : this.onDragOver} onClick={this.onPanelClicked}>
         <DockTabs panelData={panelData} onPanelDragStart={onPanelHeaderDragStart}
                   onPanelDragMove={this.onPanelHeaderDragMove} onPanelDragEnd={this.onPanelHeaderDragEnd}/>
         {isFloat ?
           [
+            <DragDropDiv key="drag-size-t" className="dock-panel-drag-size dock-panel-drag-size-t"
+                         onDragStartT={this.onPanelCornerDragT} onDragMoveT={this.onPanelCornerDragMove}
+                         onDragEndT={this.onPanelCornerDragEnd}/>,
+            <DragDropDiv key="drag-size-b" className="dock-panel-drag-size dock-panel-drag-size-b"
+                         onDragStartT={this.onPanelCornerDragB} onDragMoveT={this.onPanelCornerDragMove}
+                         onDragEndT={this.onPanelCornerDragEnd}/>,
+            <DragDropDiv key="drag-size-l" className="dock-panel-drag-size dock-panel-drag-size-l"
+                         onDragStartT={this.onPanelCornerDragL} onDragMoveT={this.onPanelCornerDragMove}
+                         onDragEndT={this.onPanelCornerDragEnd}/>,
+            <DragDropDiv key="drag-size-r" className="dock-panel-drag-size dock-panel-drag-size-r"
+                         onDragStartT={this.onPanelCornerDragR} onDragMoveT={this.onPanelCornerDragMove}
+                         onDragEndT={this.onPanelCornerDragEnd}/>,
             <DragDropDiv key="drag-size-t-l" className="dock-panel-drag-size dock-panel-drag-size-t-l"
                          onDragStartT={this.onPanelCornerDragTL} onDragMoveT={this.onPanelCornerDragMove}
                          onDragEndT={this.onPanelCornerDragEnd}/>,
@@ -326,6 +370,9 @@ export class DockPanel extends React.PureComponent<Props, State> {
   componentWillUnmount(): void {
     if (DockPanel._droppingPanel === this) {
       DockPanel.droppingPanel = null;
+    }
+    if (this._ref) {
+      this._ref.removeEventListener('pointerdown', this.onFloatPointerDown, {capture: true});
     }
     this._unmounted = true;
   }